backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
alembic==1.12.1
openpyxl==3.1.2
pandas==2.1.3
apscheduler==3.10.4
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2
email-validator==2.1.0.post1
requests==2.31.0 